Move source JSON schema test to common testing tools and add to GH source
